commit auth Edit admin

+++ app/Livewire/PartnersTable.php
@@ -34,7 +34,21 @@ final class PartnersTable extends PowerGridComponent
     public function setUp(): array
     {
         $this->showCheckBox();
+        if (auth()->user()->role === 'admin') {
+            return [
 
+                // Header::make()->showSearchInput(),
+                // Footer::make()
+                //     ->showPerPage()
+                //     ->showRecordCount(),
+                Header::make()
+                    ->showSearchInput()
+                    ->showToggleColumns(),
+                Footer::make()
+                    ->showPerPage()
+                    ->showRecordCount(),
+            ];
+        }
         return [
             Exportable::make('export')
                 ->striped()
@@ -165,6 +179,20 @@ final class PartnersTable extends PowerGridComponent
 
     public function header(): array
     {
+        if (auth()->user()->role === 'admin') {
+            return [
+                Button::add('refresh')
+                    ->slot('<div class="bg-transparent dark:bg-pg-primary-800 font-semibold py-1.5 px-3 border border-gray-300 hover:border-transparent rounded" style="border-radius:5px; background-color:white;"> <i class="fa-solid fa-rotate"></i> </div>')
+                    ->dispatch('pg:eventRefresh-default', []),
+                Button::add('add')
+                    ->slot('
+                <a href="' . route('partner.create') . '" class="bg-white font-semibold py-1.5 px-3 border border-gray-300 hover:border-gray-400 rounded inline-block">
+                    <i class="fa-solid fa-plus"></i> 
+                </a>'),
+
+            ];
+        }
+
         return [
             Button::add('refresh')
                 ->slot('<div class="bg-transparent dark:bg-pg-primary-800 font-semibold py-1.5 px-3 border border-gray-300 hover:border-transparent rounded" style="border-radius:5px; background-color:white;"> <i class="fa-solid fa-rotate"></i> </div>')
@@ -250,9 +278,11 @@ final class PartnersTable extends PowerGridComponent
     #[\Livewire\Attributes\On('bulkDelete')]
     public function bulkDelete(): void
     {
-        partner::whereIn('id', $this->checkboxValues)->delete();
-        $this->reset('checkboxValues');
-        $this->dispatch('pg:eventRefresh-default');
+        if (auth()->user()->role === 'supervisor') {
+            partner::whereIn('id', $this->checkboxValues)->delete();
+            $this->reset('checkboxValues');
+            $this->dispatch('pg:eventRefresh-default');
+        }
     }
 
     public function updatedExcelFile()
